Extract functions into seperate file
